+++ demo/routes.js
@@ -10,5 +10,14 @@ angular.module('ngtweet-test').config(['$routeProvider',
       }).
       when('/ids', {
           templateUrl: 'partials/ids.html'
+      }).
+      when('/timelines', {
+          templateUrl: 'partials/timelines.html'
+      }).
+      when('/lists', {
+          templateUrl: 'partials/lists.html'
+      }).
+      when('/searches', {
+          templateUrl: 'partials/searches.html'
       });
   }]);